requirements.txt
elevenlabs
google-generativeai
streamlit
pillow
markdown2
regex